Header file

+++ 0x15-file_io/holberton.h
@@ -1,7 +1,8 @@
 #ifndef _HOLBERTON_H_
 #define _HOLBERTON_H_
 
-long read_textfile(const char *filename, long letters);
+#include <stdlib.h>
+ssize_t read_textfile(const char *filename, size_t letters);
 int create_file(const char *filename, char *text_content);
 int append_text_to_file(const char *filename, char *text_content);
